new version

+++ include/Vector.h
@@ -14,33 +14,33 @@ class Vector {
 	public:
 		Vector();
 		Vector (const Point& endPoint);
-		Vector (float x, float y, float z = 0.0);
+		Vector (double x, double y, double z = 0.0);
 		Vector (const Vector& v);
 
-		void setXComponent (float x);
-		float getXComponent() const;
+		void setXComponent (double x);
+		double getXComponent() const;
 
-		void setYComponent (float y);
-		float getYComponent() const;
+		void setYComponent (double y);
+		double getYComponent() const;
 
-		void setZComponent (float z);
-		float getZComponent() const;
+		void setZComponent (double z);
+		double getZComponent() const;
 
 		void setEndPoint (const Point& endPoint);
 		Point getEndPoint() const;
 
-		float getMagnitude() const;
+		double getMagnitude() const;
 
 		Vector& operator= (const Vector& v);
 		friend Vector operator+ (const Vector& a, const Vector& b);
-		friend Vector operator* (float scalar, const Vector& a);
+		friend Vector operator* (double scalar, const Vector& a);
 		Vector crossProduct (const Vector& v);
 
 		Vector getUnitVector() const;
 
 		// rotates the vector around the origin
 		// positive angle rotates CCW
-		Vector rotate2D (float angle);
+		Vector rotate2D (double angle);
 
 };
